Refactor iridium_server modules for Single Responsibility Principle
- Extracted `execute_sql` and batch logic from `TdsSession` into `execution.rs`.
- Extracted handshake and login logic into `handshake.rs`.
- Refactored `tds::rpc::parser.rs` into a module directory with `types.rs`, `utils.rs`, and `parser.rs`.
- Validated via `cargo test` and `cargo check`. Core parsers in `iridium_core` remain un-split due to tight cross-module trait coupling.
